Creates/removes websocket connection on login/logout

+++ cmd/static/js/post.js
@@ -22,21 +22,26 @@ post_edit_button.on('click', function() {
 
 // Delete post modal behaviour
 const postModal = document.getElementById('modal-container-post');
-const modalPostCancelBtn = document.getElementById('delete-post-sure-no');
 let  deletePostBtn = document.querySelector('#post-deleter-button')
 if (deletePostBtn !== null) {
     deletePostBtn.addEventListener('click', function() {
         postModal.style.display = "block";
     });
 }
+let modalPostCancelBtn = document.getElementById('delete-post-sure-no');
+if (modalPostCancelBtn !== null) {
+    modalPostCancelBtn.addEventListener('click', function() {
+        postModal.style.display = 'none';
+    });
+}
+let modalPostDeleteBtn = document.getElementById('delete-post-sure-yes');
+if (modalPostDeleteBtn !== null) {
+    modalPostDeleteBtn.addEventListener('click', function() {
+        deletePost(document);
+            postModal.style.display = 'none'; 
+    });
+}
 
-modalPostCancelBtn.addEventListener('click', function() {
-    postModal.style.display = 'none';
-});
-document.getElementById('delete-post-sure-yes').addEventListener('click', function() {
-    deletePost(document);
-    postModal.style.display = 'none'; 
-});
 window.addEventListener('click', function(event) {
     if (event.target === postModal) {
         postModal.style.display = 'none';
@@ -45,23 +50,28 @@ window.addEventListener('click', function(event) {
 
 // Rate comment buttons behaviour
 let rate_post_up_button = document.getElementById('post_rate_up_button');
-rate_post_up_button.addEventListener('click', function() {
-    ratePostUp();
-})
-
+if (rate_post_up_button !== null) {
+    rate_post_up_button.addEventListener('click', function() {
+        ratePostUp();
+    });
+}
 let rate_post_down_button = document.getElementById('post_rate_down_button');
-rate_post_down_button.addEventListener('click', function() {
-    ratePostDown();
-})
+if (rate_post_down_button !== null) {
+    rate_post_down_button.addEventListener('click', function() {
+        ratePostDown();
+    });
+}
 
 let rate_post_hidden = document.getElementById('post_rating_hidden');
+if (rate_post_hidden !== null) {
 let postUserRating = rate_post_hidden.getAttribute('value');
-if (postUserRating > 0) {
-    let rate_up_button = rate_post_hidden.previousElementSibling;
-    rate_up_button.style.color = 'orange';
-} else if (postUserRating < 0) {
-    let rate_down_button = rate_post_hidden.nextElementSibling;
-    rate_down_button.style.color = 'orange';
+    if (postUserRating > 0) {
+        let rate_up_button = rate_post_hidden.previousElementSibling;
+        rate_up_button.style.color = 'orange';
+    } else if (postUserRating < 0) {
+        let rate_down_button = rate_post_hidden.nextElementSibling;
+        rate_down_button.style.color = 'orange';
+    }
 }
 
 
@@ -69,6 +79,7 @@ if (postUserRating > 0) {
 
 // Add post
 function addPost() {
+    const user_encoded = $('#user-encoded').val();
     let form = $('.post-form form')[0];
     let formData = new FormData(form)
     $.ajax({
@@ -78,7 +89,7 @@ function addPost() {
         processData: false,
         contentType: false, 
         success: function(res) {
-            window.location.href = '/'
+            window.location.href = '/user/' + user_encoded + '/feed'
         },
         error: function(err) {
             console.error("Error:", err);
